% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_0s_rows.R
\name{rm_0s_by_row}
\alias{rm_0s_by_row}
\title{To remove all-0 rows in a matrix.}
\usage{
rm_0s_by_row(x)
}
\arguments{
\item{x}{Input vector, must be a matrix.}
}
\value{
A new matrix with all-0 rows removed.
}
\description{
This is a function used to remove all-0 rows in a matrix.
}
\examples{
set.seed(123)
  x = matrix(rpois(100, 0.1), 10, 10)
  rm_0s_by_row(x)
}
